+++ tket/tests/test_Architectures.cpp
@@ -50,6 +50,8 @@ SCENARIO("Testing FullyConnected") {
       }
     }
   }
+  FullyConnected arch_named(2, "test_fc");
+  REQUIRE(arch_named.get_all_nodes_vec()[0].reg_name() == "test_fc");
 }
 
 SCENARIO("Testing RingArch") {
@@ -72,6 +74,9 @@ SCENARIO("Testing RingArch") {
     Node n2("ringNode", (i + 1) % n_nodes);
     REQUIRE(arch.edge_exists(n1, n2));
   }
+
+  RingArch arch_named(2, "test_ring");
+  REQUIRE(arch_named.get_all_nodes_vec()[0].reg_name() == "test_ring");
 }
 
 SCENARIO("Testing SquareGrid") {
@@ -102,6 +107,9 @@ SCENARIO("Testing SquareGrid") {
           1);
     }
   }
+
+  SquareGrid arch_named(2, 1, 1, "test_square_grid");
+  REQUIRE(arch_named.get_all_nodes_vec()[0].reg_name() == "test_square_grid");
 }
 
 SCENARIO("Diameters") {