+++ pytket/docs/changelog.rst
@@ -7,6 +7,8 @@ Unreleased
 Minor new features:
 
 * Update to networkx 3.
+* Add "label" argument to ``SquareGrid``, ``RingArch`` and ``FullyConnected`` 
+  ``Architecture`` classes to give custom name to constructed ``Node``.
   
 Fixes: